reorganize bus strategy header
